Implement Crafting stats (#83)
* add crafting stat handlers for start and finish events

* handle crafting revealed events

* migrate to standalone LegionStat

* hook up legions to crafting stats

* add crafting difficulty stats

* add crafting treasure stats

+++ subgraphs/bridgeworld-stats/src/helpers/number.ts
@@ -10,7 +10,7 @@ export function toPaddedString(number: i32): string {
   return number.toString();
 }
 
-export function etherToWei(ether: i32): BigInt {
+export function etherToWei(ether: f64): BigInt {
   return BigInt.fromString(
     BigDecimal.fromString(ether.toString()).times(ONE_WEI).toString()
   );
